Audio Controller voice controller test code

+++ computer/src/c/audio_controller.c
@@ -24,7 +24,7 @@
 #define DEFAULT_ADDR "127.0.0.1"
 
 int global_kill = 0;  //global program kill flag, will stop all threads if set
-int vc_pipe[2];       //voice control pipe
+int vc_pipe;          //voice control pipe
 int vc_flag;          //voice control active flag
 int vc_hold_flag;     //voice control pipe active flag
 static pid_t child;   //pid of child process
@@ -32,15 +32,11 @@ static int reciever_kill_flag; //kill flag for network reciever thread
 
 //main function for the program, listens on stdin for commands
 int main(int argc, char** argv){  
-  char input[80]; //buffer for commands
-  vc_pipe[0] = -1; //set the fd's to -1
-  vc_pipe[1] = -1;
-  vc_flag = 0;      //voice control is not active
-  vc_hold_flag = 0; //nor is the voice control pipe
-  
-  
-  signal(SIGINT, shutdown); //catch SIGINT and exit cleanly
-  
+  char input[80];   //buffer for commands
+    
+  //start the voice process before doing anything significant
+  setup_voice_control();
+     
   int period = setup_codecs(); //setup the codecs and get the period size  
   
   //handler loop, runs until program is killed
@@ -124,15 +120,16 @@ int main(int argc, char** argv){
         printf("Audio Controller: Devices shutdown\n");
       
       } else if(0 == strcmp(token, "shutdown")){
-        shutdown();
+        shutdown_prog();
+        break;
         
       //Start voice_control
       } else if(0 == strcmp(token, "voice_start")){
-        start_voice_control();
+        vc_flag = 1; //signal that threads should send data to controller
       
       //Stop voice_control
       } else if(0 == strcmp(token, "voice_stop")){
-        stop_voice_control();
+        vc_flag = 0; //signal threads to stop sending data
       
       } else {
         printf("Audio Controller: Unrecognized command: \"%s\" \n", token);  
@@ -140,59 +137,72 @@ int main(int argc, char** argv){
     } else {  //could not read from stdin
       if(!global_kill){ //if global kill is 0, it was an actual error; SIGINT otherwise
         printf("Audio Controller: Error when reading input, terminating\n");
-        shutdown();
+        shutdown_prog();
       }
       break; //break so program can exit
     }
   }
   
-  //clean up the encoder
-  destroy_codecs();
+  destroy_codecs();        //clean up the encoder
+  destroy_voice_control(); //shutdown the voice controller
   
+  printf("Audio Controller: Exiting\n");
   return 0;
 }
 
-//fork off a voice controller subprocess
-int start_voice_control(){
-  if(pipe(vc_pipe)){ //create the pipe
+//setup the pipes and create a thread for the child process
+int setup_voice_control(){
+  vc_flag = 0;      //voice control is not active
+  vc_hold_flag = 0; //nor is the voice control pipe
+
+  int pipe_fd[2];
+  if(pipe(pipe_fd)){ //create the pipes
     printf("Audio Controller: Couldn't start pipe\n");
     return -1; //return in failure
   }
   
-  //point the pipe's output to stdout
-  if(dup2(stdout, vc_pipe[1]) == -1){
-    printf("Audio Controller: Couldn't copy output\n");
+  //create copy of output file descriptor for child process
+  int out_dup;
+  if((out_dup = dup(STDOUT_FILENO)) == -1){
+    printf("Audio Controller: Couldn't dup output\n");
     return -2; //return in failure
   }
   
-  if(!(child = fork())){ //spawn the child processs
-    start_voice(vc_pipe); //throw to voice controller
+  if(!(child = fork())){  //spawn the child processs
+    close(pipe_fd[1]);    //close sender side of pipe
+    pipe_fd[1] = out_dup; //give child the duped output fd
+    start_voice(pipe_fd); //throw to voice controller
+    exit(0);              //exit child program
+  } else {                //parent process
+    close(pipe_fd[0]);    //parent closes reciever side of pipe
+    vc_pipe = pipe_fd[1]; //save the pipe's fd
   }
 
-  vc_flag = 1; //ok to start sending data
   printf("Audio Controller: Voice Controller Started\n");
+  return 0;
 }
 
 //kill the output pipe, stop 
-int stop_voice_control(){
-  vc_flag = 0;    //signal thread to stop writing data
+void destroy_voice_control(){
+  if(!vc_flag) return;  //nothing to do
+  
+  vc_flag = 0;          //signal thread to stop writing data
   while(vc_hold_flag);  //wait for pending writes to finish
-  close(fd[0]);   //close the pipe, signaling that child should close
-  sleep(TIMEOUT); //wait for child to respond
+  close(vc_pipe);    //close the pipe, signaling that child should close
+  sleep(TIMEOUT);       //wait for child to respond
   
   //check if child pid is valid, try to kill the child, check if it existed
   if(child && ESRCH != kill(child,SIGKILL)){ 
     printf("Audio Controller: Voice Controller Killed Manually\n");
-    close(fd[1]); //close its fd
   }
   child = 0; //child is no longer valid
   printf("Audio Controller: Voice Controller Shutdown\n");
 }
 
 //cleanly shutdown the program
-void shutdown(){
+void shutdown_prog(){
   //shutdown the producers, wait, then shutdown consumers
-  printf("Audio Controller: Shutting down...\n")
+  printf("Audio Controller: Shutting down...\n");
   mic_kill_flag = 1;      //assert microphone kill flag
   reciever_kill_flag = 1; //assert server kill flag
   sleep(TIMEOUT);         //wait for data to finish processing
